[MZUR-221] Move models
