+++ src/App/Policies/TestPolicy.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Policies;
+
+use Domain\Users\Models\User;
+
+class TestPolicy
+{
+    /**
+     * Create a new policy instance.
+     */
+    public function __construct()
+    {
+        //
+    }
+}
