+++ src/App/Providers/AppServiceProvider.php
@@ -4,8 +4,8 @@ namespace App\Providers;
 
 use Domain\Orders\Models\Item;
 use Domain\Items\Models\ItemProcess;
-use Domain\Items\Models\OrderItem;
 use Domain\Orders\Models\Order;
+use Domain\Orders\Models\OrderItem;
 use Domain\Orders\Models\OrderProcess;
 use Domain\Processes\Models\Process;
 use Domain\Statuses\Models\Status;
