+++ src/Domain/Orders/Models/Item.php
@@ -7,7 +7,6 @@ use Barryvdh\LaravelIdeHelper\Eloquent;
 use Database\Factories\ItemFactory;
 use Domain\Barcodes\Contracts\ScannableModel;
 use Domain\Companies\Models\Company;
-use Domain\Orders\Models\Order;
 use Domain\Warehouses\Models\StorageLocation;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
