+++ src/Domain/Orders/Models/Order.php
@@ -7,7 +7,6 @@ use Barryvdh\LaravelIdeHelper\Eloquent;
 use Database\Factories\OrderFactory;
 use Domain\Barcodes\Contracts\ScannableModel;
 use Domain\Companies\Models\Company;
-use Domain\Items\Models\OrderItem;
 use Domain\Statuses\Models\Status;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
