+++ src/Domain/Orders/Models/OrderItem.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace Domain\Items\Models;
+namespace Domain\Orders\Models;
 
-use Domain\Orders\Models\Order;
 use Domain\Processes\Models\Process;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
