+++ src/Domain/Processes/Models/Process.php
@@ -7,8 +7,7 @@ use Barryvdh\LaravelIdeHelper\Eloquent;
 use Database\Factories\ProcessFactory;
 use Domain\Barcodes\Contracts\ScannableModel;
 use Domain\Companies\Models\Company;
-use Domain\Items\Models\OrderItem;
-use Domain\Orders\Models\Order;
+use Domain\Orders\Models\OrderItem;
 use Domain\Statuses\Models\Status;
 use Domain\Warehouses\Models\Workstation;
 use Illuminate\Database\Eloquent\Builder;
